pyproject.toml
[tool.poetry]
name = "nu"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.73.0"
SQLAlchemy = "^2.0.0b3"
psycopg2 = "^2.8.6"
alembic = "^1.5.8"
uvicorn = {extras = ["standard"], version = "^0.17.4"}
asyncpg = "^0.27.0"
cryptography = "^36.0.1"
passlib = "^1.7.4"
python-jose = "^3.2.0"
broadcaster = {git = "https://github.com/sgsabbage/broadcaster.git", rev = "8e27c4e"}
promise = "^2.3"
python-dateutil = "^2.8.2"
strawberry-graphql = {extras = ["opentelemetry"], version = "^0.142.1"}
tabulate = "^0.9.0"
sqlalchemy-utils = "^0.38.3"

[tool.poetry.group.dev.dependencies]
ipython = "^8.5.0"
types-tabulate = "^0.9.0.0"
types-PyYAML = "^6.0.12.1"
pytest = "^7.0.0"
black = "^22.1.0"
mypy = "^0.982"
flake8 = "^4.0.1"
isort = "^5.9.1"
autoflake = "^1.4"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.18.3"
types-python-dateutil = "^2.8.16"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "nu.main:start"
init_db = "nu.commands:init_db"
nu = "nu.cli:cli"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "nu/test"
]
asyncio_mode = "auto"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.pyright]
include = ["nu"]
